+++ src/lib/waku_store/index.ts
@@ -13,7 +13,11 @@ import { DefaultPubSubTopic, StoreCodecs } from "../constants";
 import { selectConnection } from "../select_connection";
 import { getPeersForProtocol, selectRandomPeer } from "../select_peer";
 import { hexToBytes } from "../utils";
-import { DecryptionMethod, WakuMessage } from "../waku_message";
+import {
+  DecryptionMethod,
+  DecryptionParams,
+  WakuMessage,
+} from "../waku_message";
 
 import { HistoryRPC, PageDirection } from "./history_rpc";
 
@@ -91,7 +95,7 @@ export interface QueryOptions {
    * It can be Asymmetric Private Keys and Symmetric Keys in the same array,
    * all keys will be tried with both methods.
    */
-  decryptionKeys?: Array<Uint8Array | string>;
+  decryptionParams?: DecryptionParams[];
 }
 
 /**
@@ -176,26 +180,20 @@ export class WakuStore {
 
     if (!connection) throw "Failed to get a connection to the peer";
 
-    const decryptionKeys = Array.from(this.decryptionKeys).map(
-      ([key, { method, contentTopics }]) => {
-        return {
-          key,
-          method,
-          contentTopics,
-        };
-      }
-    );
+    let decryptionParams: DecryptionParams[] = [];
+
+    this.decryptionKeys.forEach(({ method, contentTopics }, key) => {
+      decryptionParams.push({
+        key,
+        method,
+        contentTopics,
+      });
+    });
 
     // Add the decryption keys passed to this function against the
     // content topics also passed to this function.
-    if (opts.decryptionKeys) {
-      opts.decryptionKeys.forEach((key) => {
-        decryptionKeys.push({
-          key: hexToBytes(key),
-          contentTopics: contentTopics.length ? contentTopics : undefined,
-          method: undefined,
-        });
-      });
+    if (opts.decryptionParams) {
+      decryptionParams = decryptionParams.concat(opts.decryptionParams);
     }
 
     const messages: WakuMessage[] = [];
@@ -245,7 +243,7 @@ export class WakuStore {
       const pageMessages: WakuMessage[] = [];
       await Promise.all(
         response.messages.map(async (protoMsg) => {
-          const msg = await WakuMessage.decodeProto(protoMsg, decryptionKeys);
+          const msg = await WakuMessage.decodeProto(protoMsg, decryptionParams);
 
           if (msg) {
             messages.push(msg);