+++ examples/web-chat/src/WakuContext.ts
@@ -1,9 +1,0 @@
-import { createContext, useContext } from "react";
-import { Waku } from "js-waku";
-
-export type WakuContextType = {
-  waku?: Waku;
-};
-
-export const WakuContext = createContext<WakuContextType>({ waku: undefined });
-export const useWaku = () => useContext(WakuContext);