+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- **breaking**: `DecryptionParams` may be passed when using `queryHistory` instead of just keys. 
+- **breaking**: `DecryptionParams` may be passed when using `queryHistory` instead of just keys.
+- Examples have been moved to https://github.com/waku-org/js-waku-examples.
 
 ## [0.25.0] - 2022-09-5