+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.0] - 2022-09-08
+
 ### Added
 
 - Simple connection management that selects the most recent connection for store, light push and filter requests.
@@ -20,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - doc: add missing modules.
 
-## [0.25.0] - 2022-09-5
+## [0.25.0] - 2022-09-05
 
 ### Changed
 
@@ -464,7 +466,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ReactJS Chat App example](./examples/web-chat).
 - [Typedoc Documentation](https://js-waku.wakuconnect.dev/).
 
-[Unreleased]: https://github.com/status-im/js-waku/compare/v0.25.0...HEAD
+[Unreleased]: https://github.com/status-im/js-waku/compare/v0.26.0...HEAD
+[0.26.0]: https://github.com/status-im/js-waku/compare/v0.25.0...v0.26.0
 [0.25.0]: https://github.com/status-im/js-waku/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/status-im/js-waku/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/status-im/js-waku/compare/v0.22.0...v0.23.0